requirements.txt
# this list is pruned to only contain packages needed to serve on Heroku

numpy
python-dateutil
flask
flask_limiter
tornado

# because as.db is now running off PostgreSQL
psycopg2

# needed to search on arXiv (save memory)
feedparser

# comments
pymongo

# twitter
twitter
pytz


